extract PromptList from PromptSelectField
